refactor(tests): consolidate tests_rf suite into 1-to-1 module structure
