+++ tests/tests_rf/test_binding_fsm_trace.py
