+++ tests/tests_rf/test_gateway_snapshot.py
