+++ tests/tests_rf/test_lifecycle.py
