+++ tests/tests_rf/test_topology_snapshot.py
